fastapi/requirements.txt
FastAPI==0.95.0
pymysql==1.0.2
python-dotenv==0.21.0
Jinja2==3.1.2
uvicorn==0.22.0